+++ PromoLimit/Services/MlInfoDataService.cs
@@ -19,7 +19,7 @@ namespace PromoLimit.Services
 	        using var scope = _provider.CreateScope();
             return await scope.ServiceProvider.GetRequiredService<PromoLimitDbContext>().MlInfos.AsNoTracking().ToListAsync();
         }
-        public async Task<MLInfo?> GetByUserIdAsNoTracking(int userId)
+        public async Task<MLInfo?> GetByUserIdAsNoTracking(ulong userId)
         {
 	        using var scope = _provider.CreateScope();
             return /*true switch
@@ -30,7 +30,7 @@ namespace PromoLimit.Services
             //};
         }
 
-        public async Task DeleteInfo(int userId)
+        public async Task DeleteInfo(ulong userId)
         {
 	        using var scope = _provider.CreateScope();
             var tentative = await scope.ServiceProvider.GetRequiredService<PromoLimitDbContext>().MlInfos.FirstOrDefaultAsync(x => x.UserId == userId);
